+++ src/component/header/index.js
@@ -1,12 +1,16 @@
 import style from './header.module.css';
 
-const Header = ({ title, descr}) => {
+const Header = ({ title, descr, onClickButton}) => {
+  const handlerClick = () => {
+    onClickButton && onClickButton('game');
+  }
   return (
     <header className={style.root}>
       <div className={style.forest}></div>
       <div className={style.container}>
           <h1>{title}</h1>
           <p>{descr}</p>
+          <button onClick={handlerClick}>Start Game</button>
       </div>
     </header>
   )
